CPI moved to ergo-wallet
